+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.79.0] - 2021-02-04
 ### Added
 - VIMEET-1811 - Participants can edit participants list and date of a meeting
 - VIMEET-2319 - Send webinar record to speakers